Merge (#2)
* Hardcoded Database String if no string found.

* default button for DB

* Made the default server string a const so we would only need to change it one place if we wanted to.

* DBStringHandler debugging message introduced to DatabaseSelecterForm

* Edge Case - File Cases & fast fingers. Serialisation & LinkVariable.

Fixed Edge Case: File names are sometimes not the same casing between the language files.

Fixed Edge Case: starting a proceedure too fast, completed it without doing anything.

Introduced LinkVariable to FormUtils for easier assignment of variables and forms. Might need expansion in the future. Currently works on: TextBox, NumericUpDown & CheckBox.

Included a lot of Json Serialization for Ref such that it serialises as it's value instead of as it self.